refactor: init standalone packages `@felixicaza/astro-capo` and `capo-rules`
