+++ packages/astro-capo/fixtures/base/astro.config.ts
