+++ packages/astro-capo/fixtures/base/src/layouts/Layout.astro
