+++ packages/astro-capo/fixtures/base/src/pages/index.astro
