+++ packages/astro-capo/fixtures/base/tsconfig.json
