+++ packages/astro-capo/fixtures/client-router/astro.config.ts
