+++ packages/astro-capo/fixtures/client-router/src/layouts/Layout.astro
