+++ packages/astro-capo/fixtures/client-router/src/pages/about.astro
