+++ packages/astro-capo/fixtures/client-router/src/pages/blog.astro
