+++ packages/astro-capo/fixtures/client-router/src/pages/index.astro
