+++ packages/astro-capo/fixtures/client-router/tsconfig.json
