+++ packages/capo-rules/src/constants/attributes.ts
