+++ packages/capo-rules/src/rules/asyncScriptTags.ts
