+++ packages/capo-rules/src/rules/connectionTags.ts
