+++ packages/capo-rules/src/rules/preditionTags.ts
