+++ packages/capo-rules/tests/helpers/extractHeadSequence.ts
