+++ packages/capo-rules/tsdown.config.ts
@@ -1,0 +1,3 @@
+import { felixicaza } from '@felixicaza/tsdown-config'
+
+export default felixicaza()
