+++ pnpm-workspace.yaml
@@ -1,16 +1,11 @@
 packages:
-   - tests/fixtures/*
+   - packages/*
+   - packages/astro-capo/fixtures/*
 
 allowBuilds:
    esbuild: true
    sharp: true
    simple-git-hooks: true
-
-onlyBuiltDependencies:
-   - esbuild
-   - sharp
-   - simple-git-hooks
-   - unrs-resolver
 
 trustPolicyExclude:
    - "semver@6.3.1"
